1 mise a jour ajout d'un film

+++ cinema-form-film/src/main/java/fr/laerce/cinema/dao/TmdbFilmDao.java
@@ -5,9 +5,10 @@ import fr.laerce.cinema.model.TmdbFilm;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
 public interface TmdbFilmDao extends CrudRepository<TmdbFilm, Long> {
-    public TmdbFilm findById(long id);
+    public TmdbFilm findById(BigInteger id);
 }
